Restructured a little
