+++ src/database/components/index.js
@@ -1,1 +1,2 @@
-export * from './Database';
+export * from './NavDrawer';
+export * from './RecordSearchBar';
