+++ src/sign-in/index.js
@@ -1,1 +1,2 @@
 export * from './components';
+export * from './SignIn';
